requirements.txt
langchain==0.1.12
langchain-community==0.0.28
langchain-openai==0.0.8
openai==1.14.1
tiktoken==0.6.0
chromadb==0.4.24
unstructured==0.11.8
gradio==4.22.0
python-dotenv==1.0.1
PyYAML==6.0.1
GitPython==3.1.43
langfuse==2.27.2
pypdf==4.1.0
Markdown==3.6
